modified KeptSortedSet
